Promote the experimental viewport API to the viewport package. (#1032)
* Promote experimental viewport API to the viewport package and update example app.

* Remove dependency of gestures plugin

* Animate the first data point and then use instant location updates from location component.

* Address review feedback.
