+++ app/src/main/res/values/strings.xml
@@ -156,4 +156,5 @@
     <!-- Tile requests -->
     <string name="set_tile_request_delay">Set tile request delay to 2 sec</string>
     <string name="set_tile_network_request_delay">Set tile network delay to 2 sec</string>
+    <string name="follow">Follow</string>
 </resources>
