--- plugin-viewport/src/main/kotlin/com/mapbox/maps/plugin/viewport/data/ViewportProperty.kt
@@ -1,28 +1,0 @@
-package com.mapbox.maps.plugin.viewport.data
-
-import com.mapbox.geojson.Point
-
-internal sealed class ViewportProperty<T>(var override: T?, var fallback: T) {
-
-  fun get() = override ?: fallback
-
-  class CenterProperty(override: Point?, fallback: Point) : ViewportProperty<Point>(
-    override,
-    fallback
-  )
-
-  class ZoomProperty(override: Double?, fallback: Double) : ViewportProperty<Double>(
-    override,
-    fallback
-  )
-
-  class BearingProperty(override: Double?, fallback: Double) : ViewportProperty<Double>(
-    override,
-    fallback
-  )
-
-  class PitchProperty(override: Double?, fallback: Double) : ViewportProperty<Double>(
-    override,
-    fallback
-  )
-}
