--- plugin-viewport/src/test/kotlin/com/mapbox/maps/plugin/viewport/EqualHashCodeTest.kt
@@ -1,15 +1,0 @@
-package com.mapbox.maps.plugin.viewport
-
-import com.mapbox.maps.plugin.viewport.transition.ViewportTransitionOptions
-import nl.jqno.equalsverifier.EqualsVerifier
-import org.junit.Test
-
-class EqualHashCodeTest {
-  @Test
-  fun `ViewportTransitionOptions hashCode and equals test`() {
-    val clazz = ViewportTransitionOptions::class.java
-    EqualsVerifier.forClass(clazz)
-      .usingGetClass()
-      .verify()
-  }
-}
