+++ sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/data/OverviewViewportStateOptions.kt
@@ -1,4 +1,4 @@
-package com.mapbox.maps.plugin.viewport.experimental.data
+package com.mapbox.maps.plugin.viewport.data
 
 import com.mapbox.geojson.Geometry
 import com.mapbox.maps.EdgeInsets
