--- sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/data/ViewportDataSourceUpdateObserver.kt
@@ -1,13 +1,0 @@
-package com.mapbox.maps.plugin.viewport.data
-
-/**
- * Observer that gets notified whenever [ViewportData] changes.
- */
-fun interface ViewportDataSourceUpdateObserver {
-
-  /**
-   * Called whenever [ViewportData] changes.
-   * @param viewportData latest data
-   */
-  fun viewportDataSourceUpdated(viewportData: ViewportData)
-}
