--- sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/experimental/state/ViewportState.kt
@@ -1,33 +1,0 @@
-package com.mapbox.maps.plugin.viewport.experimental.state
-
-import com.mapbox.maps.plugin.animation.Cancelable
-import com.mapbox.maps.plugin.viewport.ViewportPlugin
-
-/**
- * Definition of [ViewportPlugin] states.
- *
- * [ViewportState] is used to observe the data source associated with the state and query the camera
- * to keep updating according to current viewport data.
- */
-interface ViewportState {
-  /**
-   * Observe the new camera options produced by the [ViewportState], it can be used to get the
-   * latest state [CameraOptions] for [ViewportTransition].
-   *
-   * @param viewportStateDataObserver observer that observe new viewport data.
-   * @return a handle that cancels current observation.
-   */
-  fun observeDataSource(viewportStateDataObserver: ViewportStateDataObserver): Cancelable
-
-  /**
-   * Start updating the camera for the current [ViewportState].
-   *
-   * @return a handle that cancels the camera updates.
-   */
-  fun startUpdatingCamera(): Cancelable
-
-  /**
-   * Stop updating the camera for the current [ViewportState].
-   */
-  fun stopUpdatingCamera()
-}
