+++ sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/state/FollowingViewportState.kt
@@ -1,6 +1,6 @@
-package com.mapbox.maps.plugin.viewport.experimental.state
+package com.mapbox.maps.plugin.viewport.state
 
-import com.mapbox.maps.plugin.viewport.experimental.data.FollowingViewportStateOptions
+import com.mapbox.maps.plugin.viewport.data.FollowingViewportStateOptions
 
 /**
  * The [ViewportState] that follows user's location.
