--- sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/state/ViewportStateChangedObserver.kt
@@ -1,12 +1,0 @@
-package com.mapbox.maps.plugin.viewport.state
-
-/**
- * Observer that gets notified whenever [ViewportState] changes.
- */
-fun interface ViewportStateChangedObserver {
-  /**
-   * Called whenever [ViewportState] changes.
-   * @param viewportState current states
-   */
-  fun onViewportStateChanged(viewportState: ViewportState)
-}
