+++ sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/state/ViewportStateDataObserver.kt
@@ -1,4 +1,4 @@
-package com.mapbox.maps.plugin.viewport.experimental.state
+package com.mapbox.maps.plugin.viewport.state
 
 import com.mapbox.maps.CameraOptions
