+++ sdk-base/src/test/java/com/mapbox/maps/EqualsHashCodeTest.kt
@@ -3,12 +3,12 @@ package com.mapbox.maps
 import com.mapbox.maps.plugin.Plugin
 import com.mapbox.maps.plugin.animation.CameraAnimatorOptions
 import com.mapbox.maps.plugin.animation.MapAnimationOptions
-import com.mapbox.maps.plugin.viewport.experimental.ViewportStatus
-import com.mapbox.maps.plugin.viewport.experimental.data.DefaultViewportTransitionOptions
-import com.mapbox.maps.plugin.viewport.experimental.data.FollowingViewportStateBearing
-import com.mapbox.maps.plugin.viewport.experimental.data.FollowingViewportStateOptions
-import com.mapbox.maps.plugin.viewport.experimental.data.OverviewViewportStateOptions
-import com.mapbox.maps.plugin.viewport.experimental.data.ViewportPluginOptions
+import com.mapbox.maps.plugin.viewport.ViewportStatus
+import com.mapbox.maps.plugin.viewport.data.DefaultViewportTransitionOptions
+import com.mapbox.maps.plugin.viewport.data.FollowingViewportStateBearing
+import com.mapbox.maps.plugin.viewport.data.FollowingViewportStateOptions
+import com.mapbox.maps.plugin.viewport.data.OverviewViewportStateOptions
+import com.mapbox.maps.plugin.viewport.data.ViewportPluginOptions
 import nl.jqno.equalsverifier.EqualsVerifier
 import nl.jqno.equalsverifier.Warning
 import org.junit.Test
